feat: separate base component
